Оформил по минимуму интерфейс
