+++ src/main/resources/static/js/main.js
@@ -1,0 +1,4 @@
+
+
+@import
+"static/sass/main"
