+++ src/main/resources/static/sass/main.css
@@ -1,0 +1,14 @@
+
+.main {
+    background: cornsilk;
+    margin-left: 60px;
+}
+
+.link {
+    fill: none;
+    stroke: #bbb;
+}
+
+.container {
+    margin-left: 60rem;
+}
