+++ src/main/resources/templates/parts/common.ftl
@@ -4,7 +4,7 @@
     <head>
         <meta charset="UTF-8">
         <title>CreaterOfSocialGraph</title>
-        <link rel="stylesheet" type="../text/css" href="../css/graphStyle.css.css" >
+        <#--<link rel="stylesheet" type="text/css" href="../../static/sass/main.css">-->
 
         <meta charset="utf-8">
         <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
@@ -14,16 +14,25 @@
 
         <script src="https://cdn.jsdelivr.net/npm/vue/dist/vue.js"></script>
         <script src="https://cdn.jsdelivr.net/npm/vue-resource@1.5.1"></script>
+
+
     </head>
 
     <body>
     <script src="https://code.jquery.com/jquery-3.3.1.min.js" crossorigin="anonymous"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
+
     <#include "navbar.ftl">
-    <div class="container mt-3">
+
+    <div id="maindiv" class="container">
     <#nested>
     </div>
+
+
+    <#--  JS   -->
+    <#--<script src="../../static/js/main.js"></script>-->
+
     <!-- Optional JavaScript -->
     <!-- jQuery first, then Popper.js, then Bootstrap JS -->
     </body>
